requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
fake-useragent==0.1.11
falcon==3.0.1
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
lxml==4.6.3
PyYAML==5.4.1
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
zope.event==4.5.0
zope.interface==5.4.0
